daemon: move /v2/connections handler and tests to separate files
