+++ daemon/api.go
@@ -184,12 +184,6 @@ var (
 		POST:     changeInterfaces,
 	}
 
-	connectionsCmd = &Command{
-		Path:   "/v2/connections",
-		UserOK: true,
-		GET:    getConnections,
-	}
-
 	// TODO: allow to post assertions for UserOK? they are verified anyway
 	assertsCmd = &Command{
 		Path:   "/v2/assertions",
@@ -1728,141 +1722,6 @@ func setSnapConf(c *Command, r *http.Request, user *auth.UserState) Response {
 	st.EnsureBefore(0)
 
 	return AsyncResponse(nil, &Meta{Change: change.ID()})
-}
-
-type collectFilter struct {
-	snapName     string
-	plugSlotName string
-	ifaceName    string
-	connected    bool
-}
-
-func (c *collectFilter) plugMatches(plug *interfaces.PlugRef, connectedSlots []interfaces.SlotRef) bool {
-	for _, slot := range connectedSlots {
-		if c.slotMatches(&slot, nil) {
-			return true
-		}
-	}
-	if c.snapName != "" && plug.Snap != c.snapName {
-		return false
-	}
-	if c.plugSlotName != "" && plug.Name != c.plugSlotName {
-		return false
-	}
-	return true
-}
-
-func (c *collectFilter) slotMatches(slot *interfaces.SlotRef, connectedPlugs []interfaces.PlugRef) bool {
-	for _, plug := range connectedPlugs {
-		if c.plugMatches(&plug, nil) {
-			return true
-		}
-	}
-	if c.snapName != "" && slot.Snap != c.snapName {
-		return false
-	}
-	if c.plugSlotName != "" && slot.Name != c.plugSlotName {
-		return false
-	}
-	return true
-}
-
-func (c *collectFilter) ifaceMatches(ifaceName string) bool {
-	if c.ifaceName != "" && c.ifaceName != ifaceName {
-		return false
-	}
-	return true
-}
-
-func collectConnections(ifaceMgr *ifacestate.InterfaceManager, filter collectFilter) interfaceJSON {
-	repo := ifaceMgr.Repository()
-	ifaces := repo.Interfaces()
-
-	var ifjson interfaceJSON
-	plugConns := map[string][]interfaces.SlotRef{}
-	slotConns := map[string][]interfaces.PlugRef{}
-
-	for _, cref := range ifaces.Connections {
-		if !filter.plugMatches(&cref.PlugRef, nil) && !filter.slotMatches(&cref.SlotRef, nil) {
-			continue
-		}
-		plugRef := interfaces.PlugRef{Snap: cref.PlugRef.Snap, Name: cref.PlugRef.Name}
-		slotRef := interfaces.SlotRef{Snap: cref.SlotRef.Snap, Name: cref.SlotRef.Name}
-		plugID := plugRef.String()
-		slotID := slotRef.String()
-		plugConns[plugID] = append(plugConns[plugID], slotRef)
-		slotConns[slotID] = append(slotConns[slotID], plugRef)
-	}
-
-	for _, plug := range ifaces.Plugs {
-		plugRef := interfaces.PlugRef{Snap: plug.Snap.InstanceName(), Name: plug.Name}
-		connectedSlots, connected := plugConns[plugRef.String()]
-		if !connected && filter.connected {
-			continue
-		}
-		if !filter.ifaceMatches(plug.Interface) || !filter.plugMatches(&plugRef, connectedSlots) {
-			continue
-		}
-		var apps []string
-		for _, app := range plug.Apps {
-			apps = append(apps, app.Name)
-		}
-		pj := &plugJSON{
-			Snap:        plugRef.Snap,
-			Name:        plugRef.Name,
-			Interface:   plug.Interface,
-			Attrs:       plug.Attrs,
-			Apps:        apps,
-			Label:       plug.Label,
-			Connections: connectedSlots,
-		}
-		ifjson.Plugs = append(ifjson.Plugs, pj)
-	}
-	for _, slot := range ifaces.Slots {
-		slotRef := interfaces.SlotRef{Snap: slot.Snap.InstanceName(), Name: slot.Name}
-		connectedPlugs, connected := slotConns[slotRef.String()]
-		if !connected && filter.connected {
-			continue
-		}
-		if !filter.ifaceMatches(slot.Interface) || !filter.slotMatches(&slotRef, connectedPlugs) {
-			continue
-		}
-		var apps []string
-		for _, app := range slot.Apps {
-			apps = append(apps, app.Name)
-		}
-		sj := &slotJSON{
-			Snap:        slotRef.Snap,
-			Name:        slotRef.Name,
-			Interface:   slot.Interface,
-			Attrs:       slot.Attrs,
-			Apps:        apps,
-			Label:       slot.Label,
-			Connections: connectedPlugs,
-		}
-		ifjson.Slots = append(ifjson.Slots, sj)
-	}
-	return ifjson
-}
-
-func getConnections(c *Command, r *http.Request, user *auth.UserState) Response {
-	query := r.URL.Query()
-	snapName := query.Get("snap")
-	plugSlotName := query.Get("name")
-	ifaceName := query.Get("interface")
-	qselect := query.Get("select")
-	if qselect != "all" && qselect != "" {
-		return BadRequest("unsupported select qualifier")
-	}
-	onlyConnected := qselect == ""
-
-	ifjson := collectConnections(c.d.overlord.InterfaceManager(), collectFilter{
-		snapName:     snapName,
-		plugSlotName: plugSlotName,
-		ifaceName:    ifaceName,
-		connected:    onlyConnected,
-	})
-	return SyncResponse(ifjson, nil)
 }
 
 // interfacesConnectionsMultiplexer multiplexes to either legacy (connection) or modern behavior (interfaces).
